+++ Changes.md
@@ -1,3 +1,7 @@
+WIP  TBD
+
+ * Adding deferred.Error for helping with deferred error handling.
+
 v0.6.0  2023-08-12
 
  * Adding slices.Insert